+++ src/main/java/customermanagement/customer/controller/CustomerTemplateController.java
@@ -1,33 +1,19 @@
 package customermanagement.customer.controller;
 
-import customermanagement.customer.dto.CustomerDTO;
-import customermanagement.customer.service.ICustomerService;
+import customermanagement.customer.service.CustomerService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.List;
-
 @Controller
-@RequestMapping
 public class CustomerTemplateController {
 
-    private final ICustomerService customerService;
+    private final CustomerService customerService;
 
     @Autowired
-    public CustomerTemplateController(ICustomerService customerService) {
+    public CustomerTemplateController(CustomerService customerService) {
         this.customerService = customerService;
-    }
-
-    /**
-     * Return all customers from database
-     * @return - return with all customer
-     */
-    @GetMapping("/api/customer")
-    @ResponseBody
-    public List<CustomerDTO> findAll(){
-        return customerService.findAll();
     }
 
     /**
@@ -37,19 +23,8 @@ public class CustomerTemplateController {
      */
     @GetMapping(value = "/")
     public String index(Model model){
-        model.addAttribute("customers", findAll());
+        model.addAttribute("customers", customerService.findAll());
         return "index";
-    }
-
-    /**
-     * Return customer by ID from database
-     * @param id - customer ID
-     * @return - return customer by ID
-     */
-    @GetMapping(value = "/api/customer/{id}")
-    @ResponseBody
-    public CustomerDTO findById(@PathVariable Long id){
-        return customerService.findById(id);
     }
 
 }
