+++ src/main/resources/application.properties
@@ -7,7 +7,4 @@ spring.datasource.password=eteo
 spring.datasource.name=SecurityDB
 spring.jpa.hibernate.naming.physical-strategy=org.hibernate.boot.model.naming.PhysicalNamingStrategyStandardImpl
 
-#spring.security.user.name=eteo
-#spring.security.user.password=eteo
-
 #spring.jpa.show-sql=true
